Mouse Drag 99% Completed
I was getting a strange conflict where my mouse drag would rotate the house in the proper direction but it would also attempt to orient itself back to the original rotation. I realized after that my input settings had the mouse button mapped as an alternate button for that function so it's been removed. Now the Return function is tied to the right click.