Work on rotation
I found a way to have the rotation of the house freeze once the player holds the mouse button and keeps it still. When the mouse button is released, the last applied torque that I stored in OnMouseDrag is applied to the house. In a tactile sense, this is much less frustrating than having the player use multiple mouse drags to bring it to a stop.