First Android Build
Here's where I explain that I learned the OnMouseDrag and OnMouseUp functions work with touch controls too. Because of that, I was able to get a decent Android build right away. The only complaints I have are a need to include UI for the snap-back function as well as to quit the game.